refactor: refactor backend, update to prompts and tool descriptions, cleaned up agents file
