+++ backend/src/optcg/tools.py
@@ -1,8 +1,6 @@
 # region Imports
 from asyncio.log import logger
-from asyncio.log import logger
 import os
-from dotenv import load_dotenv
 from langchain_core.tools import tool
 from langchain_core.tools.retriever import create_retriever_tool
 from langchain_community.tools import BraveSearch, YouTubeSearchTool
@@ -15,17 +13,10 @@ import requests
 
 # Custom Imports
 from optcg.vectorstore_logic import create_or_load_vectorstore_optcg_rulebooks
-from optcg.models import CardSearchRequest, BoardState
+from optcg.schemas import CardSearchRequest
 from optcg import state
 
 api_base_url = os.getenv("API_BASE_URL", "http://localhost:8000")
-
-## Available Tools
-# - create_rulebook_retriever_tool() -- (use as function call)
-# - get_board_tool
-# - web_search_tool
-# - youtube_search_tool
-# - card_search_tool
 
 # endregion Imports
 
@@ -39,7 +30,14 @@ def create_rulebook_retriever_tool():
     rulebook_retriever_tool = create_retriever_tool(
     retriever=vectorstore.as_retriever(), # type: ignore
     name="rulebooks_retriever",
-    description="""Retrieves relevant information from the One Piece TCG rulebooks. This tool is useful for answering questions about the rules of the game, such as how to play, game setup, keywords, and tournament rules. This is a retrieval tool that will return relevant document chunks from the vectorstore, which contains the rulebooks of the One Piece TCG."""
+    description="""Retrieves relevant information from the One Piece TCG rulebooks. This tool is useful for answering questions about the rules of the game, such as how to play, game setup, keywords, and tournament rules.
+    
+    Args:
+      query (str): The query to search for in the rulebooks.
+
+    Returns:
+        A list of relevant document chunks from the rulebooks, or an error message if no relevant information is found.
+    """
     )
     return rulebook_retriever_tool
 # endregion rulebook_retriever_tool
@@ -52,7 +50,10 @@ def create_rulebook_retriever_tool():
 # Depends on the process memory or state management to store the current board state
 @tool
 def get_board_tool() -> dict:
-    """Tool that retrieves the game board state set up by the user for the One Piece TCG. Returns the current board state as a JSON. If no board state is set, returns an error message."""
+    """Retrieves the game board state set up by the user for the One Piece TCG.
+
+    Returns:
+      The current board state as a JSON object or an error message if no board state is set."""
     if state.current_board_state is None:
         logger.debug("No board state found. Returning 404.")
         return {"error": "No board state found. Please tell user to update the board state first."}
@@ -62,7 +63,7 @@ def get_board_tool() -> dict:
 # For testing purposes, can be removed later
 @tool 
 def get_board_tool_http() -> dict:
-    """Tool that retrieves the game board state set up by the user for the One Piece TCG. Returns the current board state as a JSON. If no board state is set, returns an error message."""
+    """Retrieves the game board state set up by the user for the One Piece TCG. Returns the current board state as a JSON. If no board state is set, returns an error message."""
     response = None
     try:
         response = requests.get(
